break out topic from main dao into separate
