+++ ExtendedLearning/src/main/java/com/skilldistillery/sdelp/data/SdelDao.java
@@ -13,15 +13,7 @@ public interface SdelDao {
 	
 
 	
-	List<Topic> findTopicsBySearchTerm(String keyword);
-	
-	List<Topic> getAllTopics();
-	
-	Topic getTopicById(int id);
-	
-	Topic addTopic(Topic topic);
-	
-	Topic updateTopic(Topic topic);
+
 	
 	Resource addResource(Resource resource);
